requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
awscli==1.22.80
bidict==0.21.4
boto3==1.21.25
botocore==1.24.25
cffi==1.15.0
click==8.0.4
colorama==0.4.3
cycler==0.11.0
Deprecated==1.2.13
dnspython==1.16.0
docutils==0.15.2
eventlet==0.30.2
Flask==2.0.3
Flask-Caching==1.10.1
Flask-SocketIO==5.1.1
fonttools==4.30.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
kiwisolver==1.4.0
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyasn1==0.4.8
pycparser==2.21
pymongo==4.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-librtmp==0.3.0
python-socketio==5.5.2
PyYAML==5.4.1
redis==4.2.0
rsa==4.7.2
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
